requirements.txt
pandas>=2.0,<3
numpy>=1.24,<2
matplotlib>=3.7,<4
seaborn>=0.12,<1
statsmodels>=0.14,<1
prophet>=1.1,<2
scikit-learn>=1.3,<2
PyYAML>=6.0,<7
